ui/requirements.txt
flask
pyjwt
openpyxl
nuitka